Preset trigger icon Widget width (optional)
The element that will contain your widget trigger (i.e. the HTML
element with id #noticeable-icon in this sample) is empty and has no
width set by default.

Since the Noticeable SDK but also rendering is made asynchronously, upon
page load, we don't know yet the width of the trigger icon injected in
the #noticeable-icon element. Later, when the trigger icon is injected
with a width > 0, the icon may shift other elements on your page
depending on its location.

A solution to this problem is to get the width of your trigger icon
element once loaded, and then preset the width of the #noticeable-icon
element to this value. This way, the space is pre-allocated and your
layout will remain the same after widget rendering.

Note that a width of 39px is not necessarily what you need. The value to
use depends of the customization options you set with the live
editor. Apply your configuration, render the widget and then get the
width of your trigger icon using your browser dev tools (e.g. right
click on the element, then inspect).